requirements.txt
astroid==2.2.2
grpcio==1.19.0
grpcio-tools==1.19.0
isort==4.3.9
lazy-object-proxy==1.3.1
mccabe==0.6.1
protobuf==3.7.0
pylint==2.3.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
